% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{d6Twice}
\alias{d6Twice}
\title{Roll Two 6-Sided Dice}
\usage{
d6Twice()
}
\value{
a random interger with a value between 2 and 12 inclusive. (non-uniform)
}
\description{
\code{d6Twice} returns a random interger between 2 and 12, simulating the sum
of rolling two six sided dice.
}
\details{
\code{d6Twice} utilizes the \link{dice} function to simulate the dice roll,
this function returns an interger between 1 and 6 thus \code{d6Twice} will
return a value of 2 to 12 that is not uniformly distrubited. The most common
value returned will be 7 with the least being 2 and 12.
}
\examples{
d6Twice()

}
