% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package4.R
\docType{package}
\name{Package4}
\alias{Package4}
\alias{Package4-package}
\title{Package4: A learning package for throwing dice.}
\description{
The package provides two functions \code{dice} and
\code{d6Twice}
}
\section{RpackageLearn functions}{


\link{dice} which simulates a single throw of an N sided dice

\link{d6Twice} which provides the sum of two six sided dice.
}

