% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{Dice Roll}
\usage{
dice()
}
\value{
A random interger with a value between 1 and 6 inclusive.
}
\description{
\code{dice} returns a random interger with values between 1 and 6, simulating
a roll of a single fair six sided die.
}
\details{
\code{dice} utilizes the runif function to achive the output and therefor
can be made repeatable but utilizing the \link[base]{set.seed} functionality.
While reproducability is needed at times this should only be utilized sparingly
and with caution because when used the \code{dice} becomes no longer "fair."
}
\examples{
dice()
}
