% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package4.R
\docType{package}
\name{Package4}
\alias{Package4}
\title{Package4: A learning package for throwing dice.}
\description{
The package provides three functions \code{dice}, \code{d6Twice}
and \code{DD5E}
}
\section{RpackageLearn functions}{


\link{dice} which simulates a single throw of an N sided dice. The default is
a standard six (6) sided dice. Dice must have a minuman of 2 sides (a coin).

\link{d6Twice} which provides the sum of two six sided dice.

\link{DD5E} which provides for the generation of a single character charastic
for D&D fifth edition.
}

